+++ client/scripts/views/sublayout_header_right.tsx
@@ -11,7 +11,7 @@ import NewProposalButton, {
 import { ChainInfo } from 'client/scripts/models';
 import { NotificationsMenu } from 'views/components/header/notifications_menu';
 import { InvitesMenu } from 'views/components/header/invites_menu';
-import LoginSelector from 'views/components/header/login_selector';
+import { LoginSelector } from 'views/components/header/login_selector';
 import { CWGradientButton } from './components/component_kit/cw_gradient_button';
 
 type SublayoutHeaderRightAttrs = {
@@ -50,7 +50,7 @@ export class SublayoutHeaderRight
         {app.isLoggedIn() && <NotificationsMenu />}
         {/* {app.isLoggedIn() && <InvitesMenu />} */}
         <InvitesMenu />
-        {m(LoginSelector)}
+        <LoginSelector />
       </div>
     );
   }
